break out two versions of the top-level app.js file to platform-specific versions
