+++ client/index.ios.js
@@ -1,6 +1,6 @@
 'use strict';
 
 import React, { AppRegistry } from 'react-native';
-import App from './src/app';
+import App from './src/app.ios';
 
 AppRegistry.registerComponent('trailangel', () => App);
